godiva2: Use local copies of css, images and javascript
In order to prepare for https access to tds, avoid using hardcoded paths to
http://. As some resources were not available on https the resources had to
be downloaded as local scripts. YUI is open source, license file added in
commit. As for the Reading e-Science Centre logo I have not found any
licensing terms, but it is included in order to give due credit.